Add loading.js with custom loading functionality
A new JavaScript file, loading.js, has been added which includes a host of functions related to customizing a loading state. These include an ability to start, stop, and update loading options and defaults, as well as allowing for a loading state duration with a callback. This makes it possible to specify the size, color, speed, font, message, and whether or not the loading state is fullscreen.